Redesign get_total and copy functions
Previously os.walk was called two times: when counting total number of files and copying. Now it is called only once for creating a todo list. Also there is only one copy function left, it takes this todo list as an argument. Length of this list represents total number of the files to copy.